Update directory structure
Refactor code
