+++ android/java/org/chromium/chrome/browser/shields/BraveShieldsMenuObserver.java
@@ -3,16 +3,16 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.chromium.chrome.browser.appmenu;
+package org.chromium.chrome.browser.shields;
 
 /**
  * Allows monitoring of brave shields menu actions.
  */
 public interface BraveShieldsMenuObserver {
-    /**
-     * Informs when the any of BraveShields Menu switch changes.
-     * @param isOn Whether shield is on.
-     * @param isTopShield Whether it is a top shield or not.
-     */
-    public void onMenuTopShieldsChanged(boolean isOn, boolean isTopShield);
+	/**
+	 * Informs when the any of BraveShields Menu switch changes.
+	 * @param isOn Whether shield is on.
+	 * @param isTopShield Whether it is a top shield or not.
+	 */
+	public void onMenuTopShieldsChanged(boolean isOn, boolean isTopShield);
 }
