--- android/java/res/anim/view_slide_down.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<set xmlns:android="http://schemas.android.com/apk/res/android" >
-
-<translate
-    android:duration="1000"
-    android:fromYDelta="100%"
-    android:toYDelta="0%" />
-</set>
